+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## WIP
+- add `mempoolRetryLimit` option to `postTx`
+
 ## 0.32.1 - 2020-05-31
 - update deps